refactor: organize and improve imports and exports
