+++ src/app/page.tsx
@@ -1,12 +1,10 @@
+import { CopyrightFooter, Footer } from "@/components";
 import { Section } from "@/sections";
-import Footer from "@/components/footer";
-import Hero from "@/sections/hero";
-import CopyrightFooter from "@/components/copyrightFooter";
 
 export default function Home() {
   return (
     <>
-      <Hero />
+      <Section.Hero />
       <Section.Root>
         <Section.Event id="schedule" />
         <Section.Informations id="information" />
