+++ src/sections/eventCalendar/index.tsx
@@ -1,8 +1,7 @@
 "use client";
+import { Calendar, SpecialTitle } from "@/components";
 import { motion } from "framer-motion";
-import { Calendar } from "@/components/calendar";
 import { Wrapper } from "./style";
-import { SpecialTitle } from "@/components/specialTitle";
 
 interface PropsType {
   id: string;
